GH-14844: [Java] Short circuit null checks when comparing non null field types (#15106)
Relates to #14844 

To avoid expensive checks for null values when the vectors cannot contain nulls, a flag is set when a vector is attached to indicate if a null check is needed when `compare(idx1, idx2)` is called. If it isn't then the call will immediately redirect to `compareNotNull(index1, index2)`.
* Closes: #14844

Authored-by: Mark Schreiber <mrschre@amazon.com>
Signed-off-by: David Li <li.davidm96@gmail.com>