ARROW-18195: [C++] Fix case_when produces bad data when condition has nulls (#15131)
When the condition has nulls, we were going into a branch that did not advance the `offset`. On subsequent blocks (each 64 values), the `mask` (which tells us which positions have already been set by previous cases) was always already set, so the condition was always evaluated as False regardless of its true values. So the first 64 values would be correct, but every value afterwards would skip the condition with nulls and use the next condition (or the else).

Lead-authored-by: Will Jones <willjones127@gmail.com>
Co-authored-by: Antoine Pitrou <antoine@python.org>
Signed-off-by: Antoine Pitrou <antoine@python.org>